refactor(server): extract the HTTP request router into src/core/server/request-handler.ts
Behavior-preserving relocation of the 455-line createServer callback out of
server.ts into a createRequestHandler(deps) factory. The ordered if/else dispatch
is byte-for-byte unchanged — only the boot-time closure vars (port, CONTENT_DIR,
collectOpenApiSources) are injected as deps, and the two inline require() paths
(`../settings`, `../search-registry`) are repointed for the new file location.
server.ts shrinks 723 → 268 and is now just the CLI dispatch + singleton lock +
boot orchestration + listen; it calls createServer(createRequestHandler({...})).
Dropped ~45 router-only imports from server.ts (now in request-handler.ts).

The declarative-route-table conversion the audit suggested is behavior-touching
(changes match order/semantics) — NOT taken; noted as a follow-up. This is pure
relocation.

Verified: typecheck/eslint/full suite 5124-0/binary build (stamps reverted); the
request-handler module is statically imported at server.ts top so a broken graph
would crash boot — the binary boots clean. Live HTTP routing is verified on the
dockerized rig (combined with the activation-pipeline unify, next).
